Move implementation of reduction operations to file.
